requirements.txt
--extra-index-url https://download.pytorch.org/whl/rocm5.6
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.0.0
scikit-image>=0.19.0
